requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Synthetic data generators
sdv>=1.14.0
ctgan>=0.9.0

# ML models for utility evaluation
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
# umap-learn>=0.5.0  # optional: install separately if needed (requires llvmlite)

# AWS
boto3>=1.26.0
sagemaker>=2.200,<3.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# Config and utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
